Add files via upload
Basically the same as the Legit config, with the exception FOV, & a few other minor details. You can also use this in Deathmatch OR Demo, but not in Arms Race. There are only a few weapons configured....AK-M4-DUALBB-DEAGLE-G18...The rest are set to the default config. DO NOT USE THIS FOR MM!!!